+++ docs/changelog.rst
@@ -5,6 +5,12 @@
 Changelog
 =========
 
+master
+------
+
+* new: extend WAP web service: allow loading Jinja templates from Python package,
+    check service configuration, allow running service on root path
+
 20.2.1
 ------